SampleRate: Add operator for dividing by another sample rate

+++ src/engine/enginebuffer.cpp
@@ -809,7 +809,7 @@ void EngineBuffer::processTrackLocked(
 
     double baserate = 0.0;
     if (sampleRate.isValid()) {
-        baserate = static_cast<double>(m_trackSampleRateOld) / sampleRate;
+        baserate = m_trackSampleRateOld / sampleRate;
     }
 
     // Sync requests can affect rate, so process those first.
